new file: lab-03/main.py

+++ lab-01/main.py
@@ -35,7 +35,7 @@ class Persons:
     @staticmethod
     def is_int(value):
         try:
-            float(value)
+            int(value)
             return True
         except ValueError:
             return False
